Move shared code into transaction.go.
